\name{StanfordExampleData}
\alias{63421.gpr}
\alias{hoc.gal}
\alias{DCV2.0June06.txt}
\title{Example of MEEBO GPR files and associated doping control information.}

\description{
  63421.gpr and hoc.gal are an example of hybridization on HEEBO arrays, using a
  doping-control mixture from SFGF.
  
  DCV2.0June06.txt is an example of spike-type file from SFGF.
}

\source{These data were provided by members of the Brown lab (Stanford) and the
  Stanford Shared Genomics Core Facility.}

\keyword{datasets}




